Swap DBus / file LED activation
While the DBus service works quite well on the Jolla C, it is not
working properly for Jolla users, so first try to switch the LED by
writing to a file before falling-back to DBus if it fails.

This should hopefully provide a good user experience to all Jolla users.